pyproject.toml
[tool.poetry]
name = "repl_python3_Tradingview-Telegram-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
python-telegram-bot = "^13.8.1"
selenium = "^4.0.0"
pandas = "^1.3.4"
tabulate = "^0.8.9"
python-discord = "^1.7.3"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
